requirements.txt
Django==5.0
djangorestframework==3.15.0
drf-yasg==1.21.7
PyJWT==2.8.0
setuptools==69.5.1